FRI-128 Do not remove `Authentication` from `SecurityContext` at end.
The Spring Security documentation shows that we should be able to copy the thread-local `SecurityContext` object from one `Thread` to another and keep the user context. In the case of an async job started by an API call this is not working because although we have the `SecurityContext` the `Authentication` was missing. We assume this will not lead to a memory leak because it's what the Spring documentation recommends..